+++ extensions/mymind/src/api/index.ts
@@ -4,3 +4,4 @@ export * from "./objects";
 export * from "./spaces";
 export * from "./tags";
 export * from "./thumbnails";
+export * from "./links";